stylize favourite products page
