+++ src/app.scss
@@ -26,7 +26,7 @@ body {
   width: 100%;
   max-width: 1200px;
   margin: 0 auto;
-  padding: $header-height-mobile 16px 0;
+  padding-top: $header-height-mobile;
 
   @include on-desktop {
     padding-top: $header-height-desktop;
