+++ src/pages/FavoritesPage/index.ts
@@ -1,1 +1,1 @@
-export * from './FavoritesPage';
+export * from './FavouritesPage';
